# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # COURSES
...
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      summary: Get course details
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
  /courses/{courseId}/lessons:
    get:
      operationId: getCourseLessons
      tags: [courses]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      summary: Get lessons for a course
      responses:
        "200":
          description: Lessons list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"

  # LESSONS
...
  /codes/verify:
    post:
      operationId: verifyCode
      tags: [codes]
      summary: Verify and activate an access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeVerifyInput"
      responses:
        "200":
          description: Code verified and activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeVerifyResponse"
        "400":
          description: Invalid or already used code

  # PAYMENTS
...
  /admin/payments/{paymentId}/confirm:
    post:
      operationId: confirmPayment
      tags: [admin]
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      summary: Confirm payment and issue access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentConfirmInput"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /admin/codes:
    get:
      operationId: getAdminCodes
      tags: [admin]
      summary: List all access codes
      responses:
        "200":
          description: Codes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessCode"
    post:
      operationId: createCode
      tags: [admin]
      summary: Create a new access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeCreateInput"
      responses:
        "201":
          description: Code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessCode"
  /admin/codes/{codeId}/deactivate:
    patch:
      operationId: deactivateCode
      tags: [admin]
      parameters:
        - name: codeId
          in: path
          required: true
          schema:
            type: integer
      summary: Deactivate an access code
      responses:
        "200":
          description: Code deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessCode"
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # COURSES
...
    CourseDetail:
      type: object
      required: [id, slug, title, titleAr, titleSo, description, descriptionAr, descriptionSo, language, level, price, lessonCount, duration, thumbnailUrl, lessons]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        titleAr:
          type: string
        titleSo:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        descriptionSo:
          type: string
        language:
          type: string
          enum: [english, arabic]
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        price:
          type: number
        lessonCount:
          type: integer
        duration:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        enrolledCount:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"

    # LESSONS
    Lesson:
      type: object
      required: [id, courseId, title, titleAr, titleSo, order, duration, isLocked]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        titleSo:
          type: string
        order:
          type: integer
        duration:
          type: string
        isLocked:
          type: boolean
        hasQuiz:
          type: boolean
...
    AccessCode:
      type: object
      required: [id, code, courseId, isActive, isUsed, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        isActive:
          type: boolean
        isUsed:
          type: boolean
        usedByUserId:
          type: ["integer", "null"]
        usedByEmail:
          type: ["string", "null"]
        usedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CodeVerifyInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    CodeVerifyResponse:
      type: object
      required: [success, courseId, courseName]
      properties:
        success:
          type: boolean
        courseId:
          type: integer
        courseName:
          type: string
    CodeCreateInput:
      type: object
      required: [courseId]
      properties:
        courseId:
          type: integer

    # PAYMENTS
    Payment:
      type: object
      required: [id, userId, courseId, courseName, status, whatsappNumber, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userEmail:
          type: string
        userName:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        status:
          type: string
          enum: [pending, confirmed, rejected]
        whatsappNumber:
          type: string
        notes:
          type: ["string", "null"]
        accessCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    PaymentConfirmInput:
      type: object
      required: [generateCode]
      properties:
        generateCode:
          type: boolean

    # TESTIMONIALS
...