# lib/api-zod/src/generated/types/accessCode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AccessCode {
  id: number;
  code: string;
  courseId: number;
  courseName?: string;
  isActive: boolean;
  isUsed: boolean;
  /** @nullable */
  usedByUserId?: number | null;
  /** @nullable */
  usedByEmail?: string | null;
  /** @nullable */
  usedAt?: string | null;
  createdAt: Date;
}