# lib/api-zod/src/generated/types/codeVerifyResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CodeVerifyResponse {
  success: boolean;
  courseId: number;
  courseName: string;
}