# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  titleAr: string;
  titleSo: string;
  order: number;
  duration: string;
  isLocked: boolean;
  hasQuiz?: boolean;
}